requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
sqlalchemy==1.4.25
pymssql==2.2.2
python-dotenv==0.19.2
passlib[bcrypt]==1.7.4
jose==3.3.0
celery==5.2.1
redis==4.0.2
pillow==8.4.0
torch==1.10.0
torchvision==0.11.1
jinja2==3.0.2